entities: refactor player and projectiles to take events and items to a seperate module
