+++ omg/entities/player.py
@@ -1,8 +1,10 @@
 import arcade
-from typing import Generic, List, TypeVar, Type, Iterator
+from typing import TypeVar, Type
 from omg.mechanics import movement
-from omg.entities.projectile import ProjectileFactory, ProjectileShotEvent
+from omg.entities.projectile import ProjectileFactory
 from omg.structural.observer import ObservableSprite
+from omg.entities.events import ProjectileShotEvent
+from omg.entities.items import CircularBuffer
 
 MOVEMENT_SPEED_FORWARD = 1
 MOVEMENT_SPEED_SIDE = 1
@@ -49,9 +51,7 @@ class Player(ObservableSprite):
         self.mana_regen_cooldown = 0
 
         # Skills
-        self.skills: CircularBuffer[Type[ProjectileFactory]] = (
-            CircularBuffer(N_SKILLS_MAX)
-        )
+        self.skills = SkillManager(N_SKILLS_MAX)
 
     def on_key_press(self, key, modifiers):
         """Called whenever a key is pressed."""
@@ -106,7 +106,7 @@ class Player(ObservableSprite):
 
     def add_skill(self, skill: Type[ProjectileFactory]):
         """Add skill to the player."""
-        self.skills.add(skill)
+        self.skills.add_item(skill)
 
     def set_movement_keys(self, forward, backward, left, right):
         """Bind keys to the movement logic."""
@@ -168,64 +168,7 @@ class Player(ObservableSprite):
         )
 
 
-class CircularBuffer(Generic[T]):
-    """Circular buffer to hold a determined size of items."""
+class SkillManager(CircularBuffer[ProjectileFactory]):
+    """Manages skills of an entity."""
 
-    def __init__(self, max_size):
-        self.buffer: List[T] = [None] * max_size  # Initialize with max size
-        self.max_size = max_size
-        self.current_size = 0
-        self.index = 0
-
-    def add(self, item: T):
-        """Add item to the buffer. If the buffer is full, replace an item."""
-        if self.current_size == self.max_size:
-            self.index = (self.index + 1) % self.max_size
-        else:
-            self.index = self.current_size
-        self.buffer[self.index] = item
-        if self.current_size < self.max_size:
-            self.current_size += 1
-
-    def get_current(self) -> T:
-        """Get the current item with respect to the current index."""
-        if self.current_size == 0:
-            return None
-        return self.buffer[self.index]
-
-    def get_current_index(self) -> int:
-        """Return the current index."""
-        return self.index
-
-    def get_next(self) -> T:
-        """Get next item with respect to the current index. Update the index."""
-        if self.current_size == 0:
-            return None
-        self.index = (self.index + 1) % self.current_size
-        return self.buffer[self.index]
-
-    def get_prev(self) -> T:
-        """Get previous item with respect to the current index. Update the index."""
-        if self.current_size == 0:
-            return None
-        self.index = (self.index - 1) % self.current_size
-        return self.buffer[self.index]
-
-    def set_next(self):
-        """Shift current index to the next item."""
-        if self.current_size == 0:
-            return
-        self.index = (self.index + 1) % self.current_size
-
-    def set_prev(self):
-        """Shift current index to the previous item."""
-        if self.current_size == 0:
-            return
-        self.index = (self.index - 1) % self.current_size
-
-    def __iter__(self) -> Iterator[T]:
-        """Iterate over the buffer from the start to the valid range."""
-        start = 0
-        end = self.current_size
-        for i in range(start, end):
-            yield self.buffer[i % self.max_size]
+    pass
